+++ src/python/veaf-tools/veaf_tools/commands/waypoints.py
@@ -21,6 +21,7 @@ from veaf_tools.app import (
     logger,
     t,
 )
+from veaf_tools.helpers import confirm
 
 
 @app.command(no_args_is_help=True, help=t("cmd.extract_waypoints.help"))
@@ -54,7 +55,7 @@ def extract_waypoints(
     console.print(t("cmd.extract_waypoints.title", version=VERSION))
 
     if readme:
-        if typer.confirm(t("help.confirm_doc")):
+        if confirm(t("help.confirm_doc"), unattended=True):
             md_render = Markdown(WaypointsExtractorREADME)
             console.print(md_render)
         raise typer.Exit()
@@ -123,7 +124,7 @@ def inject_waypoints(
     console.print(t("cmd.inject_waypoints.title", version=VERSION))
 
     if readme:
-        if typer.confirm(t("help.confirm_doc")):
+        if confirm(t("help.confirm_doc"), unattended=True):
             md_render = Markdown(WaypointsInjectorREADME)
             console.print(md_render)
         raise typer.Exit()